madlibs_backend/requirements.txt
diffusers
transformers
accelerate
scipy
safetensors
torch
torchvision
torchaudio
--index-url https://download.pytorch.org/whl/cu126
fastapi
uvicorn
dspy-ai
